chore: InstancedPositionVector3 receives optional entity param
This is necessary because:

If we have an attached GraphicsV2VertexController entity,
the offsets in shared InstancedIconAttributes need to be reconfigured
for the particular "slot" location of each particular entity.

This is useful if, for example, we're iterating over entities and setting
their positions using entity.position.set(). In this case, the InstancedIconAttributes
of each entity would not take into account the slot config of the particular entity
because the update() method wasn't called, which in turn calls the setConfigForLayers()
method.